Fixed some problems with the alpha compiler.
Former-commit-id: bf0cf0f485a71230573492f6309a5682966bb952

+++ src_2D/AdvectDiffuse2D/AdvectDiffuse2DQuad.cc
@@ -222,7 +222,7 @@ void AdvectDiffuse2D_Quad_Block::allocate(const int &Ni, const int &Nj, const in
  *       reconstructions more memory efficient.
  ********************************************/
 void AdvectDiffuse2D_Quad_Block::allocate_HighOrder(const int & NumberOfReconstructions,
-						    const vector<short int> & ReconstructionOrders){
+						    const vector<int> & ReconstructionOrders){
 
   bool _pseudo_inverse_allocation_(false);
   int i;
